// src/account/dto/transactionEtherHistory.dto.ts
export class TransactionEtherHistoryDto {
  blockNumber: string;
  blockHash: string;
  hash: string;
  timeStamp: string;
  nonce: string;
  transactionIndex: string;
  from: string;
  to: string;
  value: string;
  gas: string;
  gasPrice: string;
  isError: string;
  txreceipt_status: string;
  input: string;
  contractAddress: string;
  cumulativeGasUsed: string;
  gasUsed: string;
  confirmations: string;
}